requirements.txt
algoliasearch-django>=3.0,<4.0
django
djangorestframework
PyYAML
requests
django-cors-headers
psycopg2
django-dotenv